add ability to reuse npz files for generation
